% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_sample_file}
\alias{make_sample_file}
\title{Create a Sample File in BGEN/QCTOOL Format}
\usage{
make_sample_file(eid, covariates = NULL, file_out = "output.sample")
}
\arguments{
\item{eid}{Vector of sample IDs (e.g., UKB eids).}

\item{covariates}{A data.frame with the same number of rows as \code{eid}, containing additional covariates.}

\item{file_out}{Character string for the output filename.}
}
\value{
None. Writes a file to the specified path.
}
\description{
This function generates a \code{.sample} file with the required two-line header,
compatible with UKB and other genomic tools.
}
