% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_results}
\alias{merge_results}
\title{Merge a List of Results by Component}
\usage{
merge_results(res_list)
}
\arguments{
\item{res_list}{A list of named lists or vectors to be merged. All elements
should have the same names/structure.}
}
\value{
A named list where each element contains the concatenated data
  from the corresponding components of the input list.
}
\description{
This function takes a list of results (where each result is a named list or
vector with the same structure) and merges them by concatenating
corresponding components.
}
