% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{ObsG_hatY_typeIerror_step1}
\alias{ObsG_hatY_typeIerror_step1}
\title{ObsG_hatY_typeIerror_step1}
\usage{
ObsG_hatY_typeIerror_step1(mydf)
}
\arguments{
\item{mydf}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what ObsG_hatY_typeIerror_step1() does.
}
