% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{Oracle_ablation_estimate}
\alias{Oracle_ablation_estimate}
\title{Oracle_ablation_estimate}
\usage{
Oracle_ablation_estimate(mydf, independent_indices)
}
\arguments{
\item{mydf}{TODO: describe.}

\item{independent_indices}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what Oracle_ablation_estimate() does.
}
