% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_temp_dir}
\alias{reset_temp_dir}
\title{Reset and Recreate a Temporary Directory}
\usage{
reset_temp_dir(dir)
}
\arguments{
\item{dir}{Character string. The path to the directory to be reset.}
}
\description{
This function checks if a directory exists, removes it if it does,
and then creates a fresh one.
}
