% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{score_test_ObsG_single}
\alias{score_test_ObsG_single}
\title{Observed-only Score Test for a Single SNP (with Relatedness)}
\usage{
score_test_ObsG_single(G_all, step1_pars)
}
\arguments{
\item{G_all}{A numeric vector containing the genotypes for a single SNP
across all samples.}

\item{step1_pars}{A list of pre-calculated parameters from the Step 1
LMM fitting, specifically restricted to or derived from the observed
individuals (e.g., \code{obs_protein_index}, \code{invS11_res},
\code{inv_Sigma11}, \code{SX}, \code{XtSX_inv}, \code{X}).}
}
\value{
A list containing:
\itemize{
  \item{T_score_ObsG}: The score test statistic based on observed data for the SNP.
  \item{negative_log10_pval_ObsG}: P-value on the -log10 scale.
  \item{hat_beta_ObsG}: Estimated effect size (beta) using observed data.
  \item{var_hat_beta_ObsG}: Estimated variance of the effect size.
}
}
\description{
This function performs a score test for a single genetic variant using only the
samples with observed proteomic measurements. It accounts for cryptic
relatedness and population structure among the observed individuals using a
Linear Mixed Model (LMM) framework.
}
