% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{unObsG_hatY_typeIerror_step1}
\alias{unObsG_hatY_typeIerror_step1}
\title{unObsG_hatY_typeIerror_step1}
\usage{
unObsG_hatY_typeIerror_step1(mydf)
}
\arguments{
\item{mydf}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what unObsG_hatY_typeIerror_step1() does.
}
