% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{block_matrix}
\alias{block_matrix}
\title{Construct a Block Matrix from Four Quadrants}
\usage{
block_matrix(tl, tr, bl, br)
}
\arguments{
\item{tl}{The top-left matrix block.}

\item{tr}{The top-right matrix block.}

\item{bl}{The bottom-left matrix block.}

\item{br}{The bottom-right matrix block.}
}
\value{
A single matrix formed by binding the four blocks together.
}
\description{
This function creates a large matrix by combining four individual matrices
representing the top-left, top-right, bottom-left, and bottom-right blocks.
}
