% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DGP_GRMr_gmodel_step3_rho}
\alias{DGP_GRMr_gmodel_step3_rho}
\title{Data Generation Process (DGP) for Correlated Samples (Mixed Model)}
\usage{
DGP_GRMr_gmodel_step3_rho(
  n_obs,
  miss,
  tauT = 0.7,
  sigmaT = 0.7,
  pve_x = 0,
  pve_g = 1e-04,
  Genotype_each,
  pre_vars_step2,
  GRM
)
}
\arguments{
\item{n_obs}{Number of individuals with observed phenotypes.}

\item{miss}{Proportion of missingness for the target phenotype.}

\item{tauT}{Variance component for the random effect in the target trait.}

\item{sigmaT}{Residual variance component for the target trait.}

\item{pve_x}{Proportion of variance explained by covariates.}

\item{pve_g}{Proportion of variance explained by the specific SNP (Effect Size).}

\item{Genotype_each}{A vector of genotypes for all samples (typically generated
via \code{generate_genotypes_rho}).}

\item{pre_vars_step2}{A list containing pre-calculated covariance matrices
(e.g., \code{chol_Sigma22}, \code{L_cond}, \code{Sigma12_Sigma22inv}) from Step 2.}

\item{GRM}{The full Genetic Relatedness Matrix corresponding to the samples.}
}
\value{
A list containing the simulated dataset:
\itemize{
  \item{\code{G_all}}: The input genotype vector.
  \item{\code{X_all}}: Simulated independent normal covariates.
  \item{\code{S}}: Synthetic/Surrogate phenotypes for all samples.
  \item{\code{Y}}: Oracle (fully observed) target phenotypes.
  \item{\code{Y_obs}}: Partially observed target phenotypes (with NAs).
  \item{\code{GRM}}: The Genetic Relatedness Matrix used.
}
}
\description{
This function generates phenotypes (Target Y and Surrogate S) conditioned on
a pre-existing genotype matrix (e.g., simulated with specific family
correlations) and a Genetic Relatedness Matrix (GRM). It utilizes
pre-calculated conditional distributions to ensure that the simulated traits
maintain the intended variance components and genetic architecture.
}
