% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_inv_block}
\alias{matrix_inv_block}
\title{Parallel Block-wise Matrix Inversion}
\usage{
matrix_inv_block(wait_matrix, threshold = 1000)
}
\arguments{
\item{wait_matrix}{A large square matrix (typically a sparse \code{dgCMatrix}) to be inverted.}

\item{threshold}{Integer. The maximum number of rows/columns for a merged block
to optimize memory and computation. Default is 1000.}
}
\value{
A sparse block-diagonal matrix representing the inverse of the input matrix.
}
\description{
This function performs an efficient inversion of a large sparse matrix by
partitioning it into independent blocks. It merges smaller blocks to
optimize parallel processing and uses Cholesky decomposition for
numerical stability.
}
