% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{ObsG_typeIerror_step1_lm}
\alias{ObsG_typeIerror_step1_lm}
\title{ObsG_typeIerror_step1_lm}
\usage{
ObsG_typeIerror_step1_lm(mydf)
}
\arguments{
\item{mydf}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what ObsG_typeIerror_step1_lm() does.
}
