% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{reset_temp_dir}
\alias{reset_temp_dir}
\title{reset_temp_dir}
\usage{
reset_temp_dir(dir)
}
\arguments{
\item{dir}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what reset_temp_dir() does.
}
