% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{score_test_ObsG_lm_multiply}
\alias{score_test_ObsG_lm_multiply}
\title{score_test_ObsG_lm_multiply}
\usage{
score_test_ObsG_lm_multiply(g_matrix, step1_pars)
}
\arguments{
\item{g_matrix}{TODO: describe.}

\item{step1_pars}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what score_test_ObsG_lm_multiply() does.
}
