% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{score_test_ObsG_multiply}
\alias{score_test_ObsG_multiply}
\title{Score Test for Multiple SNPs using the Observed-only Method (with Relatedness)}
\usage{
score_test_ObsG_multiply(g_matrix, step1_pars)
}
\arguments{
\item{g_matrix}{A numeric genotype matrix (n_total x m).}

\item{step1_pars}{A list of pre-calculated parameters from the Step 1 model
fitting, specifically restricted to or derived from the observed individuals
(e.g., \code{obs_protein_index}, \code{invS11_res}, \code{inv_Sigma11},
\code{SX}, \code{XtSX_inv}).}
}
\value{
A list containing:
\itemize{
  \item{T_score_ObsG}: The score test statistics based on observed data.
  \item{negative_log10_pval_ObsG}: P-values on the -log10 scale.
  \item{hat_beta_ObsG}: Estimated effect sizes (beta) using observed data.
  \item{var_hat_beta_ObsG}: Estimated variances of the effect sizes.
}
}
\description{
This function performs a score test for a batch of SNPs using only the samples
with observed proteomic measurements. It accounts for cryptic relatedness
and population structure among the observed individuals using a Linear
Mixed Model (LMM) framework.
}
