% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynPALM-package.R
\docType{package}
\name{SynPALM-package}
\alias{SynPALM}
\alias{SynPALM-package}
\title{SynPALM: Synthetic Phenotype Assisted Linear Mixed Models}
\description{
Robust and computationally scalable inference for proteome-wide GWAS when the
target phenotype is only partially observed. SynPALM jointly analyses observed
measurements and complete synthetic (machine-learning predicted) phenotypes
while accounting for cryptic relatedness and population structure via linear
mixed models.
}
\author{
\strong{Maintainer}: Haoyu Yang \email{haoyuyang@hsph.harvard.edu}

}
\keyword{internal}
