% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synpalm_functions.R
\name{SynSurrG_eachG_step1}
\alias{SynSurrG_eachG_step1}
\title{SynSurrG_eachG_step1}
\usage{
SynSurrG_eachG_step1(mydf, G)
}
\arguments{
\item{mydf}{TODO: describe.}

\item{G}{TODO: describe.}
}
\value{
TODO: describe the return value.
}
\description{
TODO: one line describing what SynSurrG_eachG_step1() does.
}
